+++ app/api/v1/models/__init__.py
@@ -1,3 +1,5 @@
+from .ballot import *
 from .base import *
 from .guardian import *
 from .key import *
+from .tally import *
