+++ app/api/v1/models/ballot.py
@@ -1,0 +1,9 @@
+from typing import Any
+
+from .base import Base
+
+
+class AcceptBallotRequest(Base):
+    ballot: Any
+    description: Any
+    context: Any
