+++ app/api/v1/models/tally.py
@@ -1,0 +1,13 @@
+from typing import Any, List
+
+from .base import Base
+
+
+class StartTallyRequest(Base):
+    ballots: List[Any]
+    description: Any
+    context: Any
+
+
+class AppendTallyRequest(StartTallyRequest):
+    encrypted_tally: Any
